API Integrator/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "api-integrator"
version = "0.1.0"
description = "AI-powered API orchestrator & HuggingFace model hub agent using Claude Agent SDK"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "typer[all]>=0.12.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
]

[project.scripts]
api-integrator = "api_integrator.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/api_integrator"]
